feat: Networks mock
Separating mockNetworks into its own file for usage by other components if needed.
