+++ src/app/__tests__/Proposals.test.tsx
@@ -2,7 +2,8 @@ import React from "react"
 import Proposals from "pages/gov/Proposals"
 import { render } from "@testing-library/react"
 import { MemoryRouter } from "react-router-dom"
-import { mockProposals, mockNetworks } from "./__mocks__/Proposals.mock"
+import { mockProposals } from "./__mocks__/Proposals.mock"
+import { mockNetworks } from "./__mocks__/Networks.mock"
 import { useQuery } from "react-query"
 import { RecoilRoot } from "recoil"
