+++ src/app/__tests__/__mocks__/Proposals.mock.tsx
@@ -148,136 +148,4 @@ const mockProposals = [
   },
 ]
 
-const mockNetworks = {
-  "atreides-1": {
-    baseAsset: "uatr",
-    chainID: "atreides-1",
-    coinType: "118",
-    gasAdjustment: 2.5,
-    gasPrices: {
-      uatr: 0,
-    },
-    ibc: {
-      fromTerra: "channel-226",
-      ics: {
-        contract:
-          "atreides1yyca08xqdgvjz0psg56z67ejh9xms6l436u8y58m82npdqqhmmtqmxgcku",
-        fromTerra: "channel-230",
-        toTerra: "channel-1",
-      },
-      icsFromTerra: {
-        contract:
-          "terra12havy0g3svy6jlnc0je3z4f0lujhkp8h79ux5myhy4ujufeczpzsr5w7lz",
-        fromTerra: "channel-233",
-        toTerra: "channel-2",
-      },
-      toTerra: "channel-0",
-    },
-    icon: "https://station-assets.terra.money/img/chains/Atreides.png",
-    lcd: "https://atreides.terra.dev:1317",
-    name: "Atreides",
-    prefix: "atreides",
-  },
-  "corrino-1": {
-    baseAsset: "ucor",
-    chainID: "corrino-1",
-    coinType: "118",
-    gasAdjustment: 2.5,
-    gasPrices: {
-      ucor: 0,
-    },
-    ibc: {
-      fromTerra: "channel-225",
-      ics: {
-        contract:
-          "corrino1yyca08xqdgvjz0psg56z67ejh9xms6l436u8y58m82npdqqhmmtqmxgcku",
-        fromTerra: "channel-231",
-        toTerra: "channel-1",
-      },
-      icsFromTerra: {
-        contract:
-          "terra12havy0g3svy6jlnc0je3z4f0lujhkp8h79ux5myhy4ujufeczpzsr5w7lz",
-        fromTerra: "channel-235",
-        toTerra: "channel-2",
-      },
-      toTerra: "channel-0",
-    },
-    icon: "https://station-assets.terra.money/img/chains/Corrino.png",
-    lcd: "https://corrino.terra.dev:1317",
-    name: "Corrino",
-    prefix: "corrino",
-  },
-  "harkonnen-1": {
-    baseAsset: "uhar",
-    chainID: "harkonnen-1",
-    coinType: "118",
-    gasAdjustment: 2.5,
-    gasPrices: {
-      uhar: 0,
-    },
-    ibc: {
-      fromTerra: "channel-228",
-      ics: {
-        contract:
-          "harkonnen1yyca08xqdgvjz0psg56z67ejh9xms6l436u8y58m82npdqqhmmtqswjsmm",
-        fromTerra: "channel-232",
-        toTerra: "channel-1",
-      },
-      icsFromTerra: {
-        contract:
-          "terra12havy0g3svy6jlnc0je3z4f0lujhkp8h79ux5myhy4ujufeczpzsr5w7lz",
-        fromTerra: "channel-236",
-        toTerra: "channel-2",
-      },
-      toTerra: "channel-0",
-    },
-    icon: "https://station-assets.terra.money/img/chains/Harkonnen.png",
-    lcd: "https://harkonnen.terra.dev:1317",
-    name: "Harkonnen",
-    prefix: "harkonnen",
-  },
-  "ordos-1": {
-    baseAsset: "uord",
-    chainID: "ordos-1",
-    coinType: "118",
-    gasAdjustment: 2.5,
-    gasPrices: {
-      uord: 0,
-    },
-    ibc: {
-      fromTerra: "channel-227",
-      ics: {
-        contract:
-          "ordos1yyca08xqdgvjz0psg56z67ejh9xms6l436u8y58m82npdqqhmmtq56pc9e",
-        fromTerra: "channel-229",
-        toTerra: "channel-1",
-      },
-      icsFromTerra: {
-        contract:
-          "terra12havy0g3svy6jlnc0je3z4f0lujhkp8h79ux5myhy4ujufeczpzsr5w7lz",
-        fromTerra: "channel-234",
-        toTerra: "channel-2",
-      },
-      toTerra: "channel-0",
-    },
-    icon: "https://station-assets.terra.money/img/chains/Ordos.png",
-    lcd: "https://ordos.terra.dev:1317",
-    name: "Ordos",
-    prefix: "ordos",
-  },
-  "pisco-1": {
-    baseAsset: "uluna",
-    chainID: "pisco-1",
-    coinType: "330",
-    gasAdjustment: 3.5,
-    gasPrices: {
-      uluna: 0.015,
-    },
-    icon: "https://station-assets.terra.money/img/chains/Terra.svg",
-    lcd: "https://pisco-lcd.terra.dev",
-    name: "Terra",
-    prefix: "terra",
-  },
-}
-
-export { mockProposals, mockNetworks }
+export { mockProposals }
